Fix calculation of T, P, and O3 in rrtmg_rad_transfer_mod.F
Updated the code in DO_RRTMG_RAD_TRANSFER (GeosCore/rrtmg_rad_transfer_mod.F)
so that we calculate T_CTM, P_CTM, and O3_CTM and pass them to SET_PROF_O3
in a manner that is consistent with the code in fast_jx_mod.F. Also moved
and fixed calculation of O3_CTM so that it is consistent with fast_jx_mod.F.
These changed fix an array-out-of-bounds error that was occurring in array
ABSB in GeosRad/rrtmg_lw_taumol.f90.

Adding the new explicit arrays T_CTM and P_CTM also prevents array
temporaries from being created for these fields during the call to
SET_PROF_O3.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>