Now add NO_BPCH C-preprocessor switch to block out old diagnostics
We now can use #if !defined( NO_BPCH ) to block out the old diagnostic
arrays that get sent to the bpch file.  We are going to be defining
new HEMCO-based containers for each GEOS-Chem diagnostic, so that the
data may be sent to netCDF files.  This renders the existing diagnostic
arrays obsolete.

As each diagnostic is converted to the new structure, we will block
out the old diagnostic arrays.  For a transition period you will be
able to use both the netCDF and bpch diagnostics.  But if you wanted
to not use the bpch diagnostic arrays you can block those out by
compiling with NO_BPCH=y.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>