Replace State_Met%PMID w/ GET_PCENTER routine call. Declare OpenMP private vars.
State_Met%PMID [Pa] is never set in GEOS-Chem so it is not used for calculating moist air density in the AIRQNT routine. Instead, GET_PCENTER, a routine defined in GeosUtil/pressure_mod.F, is used to retrieve the pressue at the center of each grid cell [hPa]. The formulas to calculate moist air density are adjusted as needed to account for the different pressure units. In addition, several variables are added to the list of declared private variables for OpenMP parallelization and the AIRQNT is generally cleaned up for readability.