Now revert to reading the lightning CDF table from a file
It turns out that inlining the Ott et al lightning CDF table
via an include file drastically slows down the compilation time.
The module lightning_cdf_mod.F90 has to always be rebuilt due to
the include.

We have now restored the reading of the CDF table from an ASCII
file in the HEMCO/LIGHTNOX/v2014-07 directory.

For now, we only read the ASCII file if am_I_Root = .TRUE.  This
is always .TRUE. for the "GEOS-Chem classic" simulation.  In a later
commit, we will have to MPI_BROADCAST out the PROFILE array to the
other CPUs.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>