Restore CH4 biogenic emissions to ND46 diagnostic
In the pre-HEMCO code, we included CH4 biogenic emissions in the ND46
diagnostic output. We now calculate CH4 biogenic emissions by obtaining
the CH4_WETLAND, CH4_OTHERNATUR, and CH4_SOILABSORB diagnostics from
HEMCO and combining. Note that we add CH4_WETLAND and CH4_OTHERNATUR, but
subtract CH4_SOILABSORB because it is a sink.

Also fixed unit conversion for the SOA biogenic emissions in the ND07
diagnostic. BIOGENIC_xxx emissions from HEMCO are in units of kgC/m2/s and
need to be converted to kgC for ND07.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>