Minor fix to write CH4 diagnostics to bpch file
Needed to add call to BPCH2 to make sure the ND46 CH4 biogenic emissions
calculated in diag3.F are written out to the bpch file.

NOTE: BIOGSRCE (ND46) for CH4 will be zeroes for all simulations except
the UCX and CH4 simulations. For UCX, turning on CH4 emissions is optional
(set in input.geos). If LCH4CHEMIS is set to false, as is the default, all
CH4 emissions will be zero.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>