Bug fix: Always allocate the ZERO_DVEL array in pops_mod.F
Dry deposition is now applied in mixing_mod.F90.  The calls to
CHEM_POPGP in GeosCore/pops_mod.F were modified to pass ZERO_DVEL
as an argument.  But ZERO_DVEL was not always allocated.  The
quick fix is to always allocate the ZERO_DVEL array within
pops_mod.F by removing it from an IF block.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>