Bug fix for O3_CTM in Fast-JX and RRTMG code
The SET_PROF and SET_PROF_O3 routines expect model O3 in units molec/cm3,
so we need to set O3_CTM from the CSPEC array, not State_Chm%Tracers (kg).
In an earlier fix (SHA1 ID 7feaa42), Sebastian Eastham said O3_CTM should
only be defined for levels 1:GET_CHEMGRID_LEVEL(I,J) because O3 is not
defined above the chemistry grid level. Here, we fix O3_CTM in
fast_jx_mod.F and rrtmg_rad_transfer_mod.F so that we set it to CSPEC in
a loop over 1 to LCHEM. That way, we properly pass the column O3 to the
SET_PROF and SET_PROF_O3 routines.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>