Fix an apparent optimization error in GeosUtil/grid_mod.F90
Under some circumstances, the YSIN array in grid_mod.F90 was
being returned as all zeroes.  We attribute this to a possible
optmization error.  The error would normally show up in the
geos5_4x5_fullchem simulation when DEBUG=n.

The error was apparently caused by an array element being passed
to the SIN function (i.e. SIN( YDGR(I,J,L ) ).  Rewriting the code
so that a scalar is passed to the SIN function, i.e.:

  YEDGE_VAL   = YDGR(I,J,L)
  YSIN_VAL    = SIN( YEDGE_VAL )
  YSIN(I,J,L) = YSIN_VAL

fixes the problem.

Our recommendation would be to pass a scalar argument to each
of the trigonometric functions going forward.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>