Bug fix: Avoid parallelization error in the ND21 diagnostic for dust
We need to set the LINTERP logical variable as !OMP+PRIVATE in routines
RDUST_ONLINE and RDUST_OFFLINE. Otherwise, we get differences in the ND21
tracer #4 and #5 diagnostic output when OMP=yes and OMP=no.

Also changed all instances of (LINTERP .EQ. .FALSE.) to ( .not. LINTERP)
to avoid potential compilation errors when comparing logical values.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>