Bug fixes for ND56 lightning diagnostics
The lightning NOx diagnostics weren't getting properly written out to the
bpch file. To resolve this, we have changed the AutoFill flag passed to
routine DIAG2BPCH to -1 so that all HEMCO diagnostics are searched.

The ND56 diagnostic has also been updated to save out lightning cloud top
heights. This diagnostic was added in HEMCO and should be written to the
ctm.bpch file in DIAG3 to avoid having it written to the HEMCO restart
file.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>