Bug fix: Deactivate parallel loop in dust_mod.F for ND21 diag
For some reason, the updates made to dust_mod.F in v10-01i (in
the parallel loop where ND21 dust diagnostics are computed) cause
differences between single and multi-processor runs.  An initial
investigation could not find the reason why this difference was
occurring.

The cheap solution is to comment out the !$OMP parallel loop
directives in the loop where this occurs.  The ND21 diagnostic is
slated to be replaced by netCDF output in the near future.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>