Prevent div-by-zero in aerosol_mod.F
We needed to add an error check to prevent a div-by-zero condition
in aerosol_mod.F.  In the UCX simulations, the SO4_NH4_NIT array
can be zero in the stratosphere.  This was put into a denominator
of a couple of expressions.  We now just set the result to zero
if SO4_NH4_NIT is zero, thus avoiding the error.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>