Minor bug fixes in fast_jx_mod.F
In GeosCore/fast_jx_mod.F, we added some missing variables (L,
LCHEM, N, IOPT) to the !$OMP+PRIVATE clause.  We also needed to
add a test for JLOOP > 0 in order to avoid an out-of-bounds
error.

Also, in the Makefile_header.mk, we now export the RRTMG variable
so that it can be used in the GeosCore/Makefile.  RRTMG is now
set to yes if it passes the regexp.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>