Modifications to input_mod.F and input.geos for CO2 update and HEMCO
Due to the implementation of HEMCO, it is no longer useful to have options for various fluxes of a specific field in input.geos.
This modification simplifies input.geos so that one only needs to select if an emissions is on/off.
Options for the emission are found in the HEMCO config file.

Signed-off-by: Ray Nassar <ray.nassar@ec.gc.ca> 2015-Apr-14