Removed obsolete AIRQNT_FULLGRID routine from GeosCore/dao_mod.F
The AIRQNT_FULLGRID routine was originally used to regrid 72-level
OH fields for the specialty simulations.  But this is now handled
via HEMCO.  Therefore, AIRQNT_FULLGRID is now obsolete, so we can
remove it.

Also -- in the version following v10-01, we will be updating the
way in which air mass quantities are computed.  So we would not be
able to keep AIRQNT_FULLGRID due to this.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>