Fixed out-of-bounds errors in rrtmg_rad_transfer_mod.F
In the original code for RRTMG (based on v9-01-03), some arrays
like CLDF and AIRDEN used (L,I,J) ordering.  But in v10-01, these
are now members of the State_Met object.

All members of State_Met now use (I,J,L) ordering, in order to
be compatible with the NASA GEOS-5 GCM.  Therefore the old
ordering was causing an array-out-of-bounds error for the
CLDF and AIRDEN fields.  We now use State_Met%CLDF(I,J,L)
and State_Met%AIRDEN(I,J,L) and that fixes the problem.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>