Bug fix in I3 field interpolation: fraction of timestep elapsed exceeds 1.
TM and TC2, defined as fraction of 6h timestep elapsed at mid-point and
end of this dynamic timestep, need to be reset after the 3hr mark for
the case of I3 fields. Fractions are correct for dynamic timesteps
0 to 2.75 but TM exceeds 1.0 for dynamic timesteps 3.0 to 5.75.

INTERP routine in dao_mod now contains a check for TM > 1.0 if GEOS-FP
is defined. If TM > 1.0, then TM and TC2 are rest to TM-1 and TC2-1.
This results in smooth interpolation of I3 fields across all dynamic
timesteps. This bug affects State_Met variables T, SPHU, and PSC2 of
GEOS-FP only.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>