Fix TOMAS bug in wetscav_mod where routines are passed v/v not kg.
Wetscav_mod routines WASHOUT and COMPUTE_F are both passed arg State_Chm
for use if TOMAS is defined. The TOMAS routines that receive State_Chm
expect tracer units in kg. However, if WASHOUT and COMPUTE_F are called
from convection, tracer units are in v/v.

This issue is now fixed by converting STT from v/v to kg within the TOMAS
ifdef blocks of WASHOUT and COMPUTE_F and then back to v/v at the end of
the blocks. The conversion only occurs if STT units are currently in v/v.
This is determined by taking the average value of STT tracer 1 at level 1
and thresholding it again 0.1. If less than 0.1, units are considered to
be v/v and the conversion to kg occurs. A flag is then set indicating the
conversion took place (UNITCHANGE) which later serves as an indicator
that units must be converted again, back to v/v.

Also in this commit are minor updates to convection_mod comments regarding
wetscav_mod allocatable array H2O2s and SO2s to be clear that units are
v/v and that the arrays are from wetscav_mod (not sulfate_mod).

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>