Test if GC_F_INCLUDE, GC_F_BINenv vars are defined before using them
We have made a minor enhancement to the Makefile_header.mk.  From
a prior commit, the GC_F_INCLUDE and GC_F_BIN variables were
introduced so that you could point to a separate netcdf-fortran
directory path.  But the great majority of GC users probably won't
need to do this.

We now revert to using GC_BIN if GC_F_BIN is not found, and
GC_INCLUDE if GC_F_INCLUDE is not found.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>