Remove more variables that were made obsolete by HEMCO
Now that EMISSIONS are handled by HEMCO, we have begun to remove
variables that were still hanging over from the old legacy
emissions code.  It is important to remove these variables,
in order to ensure that we are not allocating memory to arrays
that are not used, etc.

We have also started to change the comments in the code to
reference "GEOS-Chem HP" and "GEOS-Chem CLASSIC", which are now
the approved terms.

Also added some cosmetic changes to Headers/CMN_SIZE_mod.F, to
line up text on the same column.  We also now set DISIZE and
DJSIZE to dummy values if we are using GEOS-Chem HP.

Bob Y. performed a diff test on these modifications w/r/t the
previous commit, and obtained identical results.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>