Renamed KPP/standard to KPP/benchmark
We have renamed the KPP/standard directory to KPP/benchmark,
in order to denote that the benchmark simulation is now the
default simulation.

Also added extra logic in the Makefile_header.mk to properly
select the values of CHEM and UCX.  The default settings
are UCX=y and CHEM=benchmark.

Also renamed KPP/fullchem to KPP/NOx_Ox_HC_Br for clarity.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>