Add doc for timeseries diagnostics now archived by HEMCO
HEMCO now archives several diagnostic quantities that used to
be saved out in the ND49, ND50, and ND51 timeseries output.

The equivalent outputs can be saved to netCDF file by specifying
some extra input in the HEMCO_Config.rc and DiagnFile.rc files.
We have added a note that tells users how to do this (and also
points them to the HEMCO user's guide on the wiki).

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>