Updated Makefiles to remove module, libraries, and executables
Each subdirectory Makefile now removes the *.mod, *.a, and *.x
files that it creates.  This step was necessary in order to
implement the "make-realclean-except-RRTMG" option.

If you now type one of these commands:

  make realclean RRTMG_NO_CLEAN=y
  make realclean RRTMG_NOCLEAN=y

Then GEOS-Chem will remove everything except *.mod and *.a files
pertaining to the RRTMG simulation.  This should hopefully help
to reduce the time it takes to recompile.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>