Bug fix: also remove mersennetwister.mod when making clean
The RRTMG code has a file named "mcica_random_numbers.F90" that
contains a module of a different name "MersenneTwister".  When
we make clean, we weren't removing the mersennetwister.mod
from the modules directory.

The quick fix is to just manually add this to the modules list,
so that it will be included in the make clean command.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>