Bug fix: Don't automatically set UCX=y in most cases
In the Makefile_header.mk, we were setting the value of UCX for
the various permutations of CHEM=benchmark, CHEM=SOA, CHEM=UCX,
and CHEM=tropchem.  This was causing conflicts when manually
compiling GEOS-Chem outside of the unit tester run dirs.

The solution is to not set UCX=y unless CHEM=UCX.  Then it makes
sense to set it.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>