Now deposit Hg2, HgP to snowpack even if LDYNOCEAN=F
We were getting different values of SNOW_HG (i.e. Hg deposited
to the snowpack) when the dynamic ocean was turned on (i.e.
LDYNOCEAN=T) w/r/t when it was turned off (LDYNOCEAN=F).

To correct this, we now add wet-deposited Hg2 and HgP to the
snowpack regardless of the setting of LDYNOCEAN.  Modified
a couple of locations in convection_mod.F and wetscav_mod.F.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>