Remove ND44 code orphaned by HEMCO & cleanup code accordingly
Prior to HEMCO, the ND44 drydep diagnostic was archived in
several routines, mostly in sulfate_mod.F and carbon_mod.F.
But now, all drydep is applied in mixing_mod.F.  Therefore,
we can remove many obsolete variables from these areas of
the code.

We can also remove some IF blocks where drydep frequencies
are applied.  This allows us

Also:
(1) In sulfate_mod.F, we have removed CHEM_NH3 and CHEM_NH4,
    as these routines were not being called anymore.
(2) In seasalt_mod.F, we have removed DRY_DEPOSITION, which
    was not getting called any more
(3) We have updated the units of ND44 for the Rn-Pb-Be simulation.
    These are now archived in molec/cm2/s instead of kg/s.

NOTE: We have left the AD44 array archival intact in a few
areas that were not touched by Christoph Keller when he made
the HEMCO updates.  For example, gravitational settling is
not covered by mixing_mod.F90, so we leave that intact.

Users of specialty simulations should verify that dry-deposited
species are not being double-counted into AD44.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>