Use separate linker commands for GEOS-Chem and HEMCO standalone
Fixed a bug in Makefile_header.mk in which the HEMCO standalone
linker command was trying to link to the RRTMG directory.  This
caused the link command to die with an error.

The fix is to create separate linker commands for the GEOS-Chem
and HEMCO standalone executables so that they don't interfere
with each other.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>