Bug fix in tagged_ox_mod.F: PP does not have to be held PRIVATE
When we removed the dry deposition code from CHEM_TAGGED_OX
(in GeosCore/tagged_ox_mod.F), we had set PP as !$OMP+PRIVATE.
But because this routine has 4 dimensions it does not have
to be !$OMP+PRIVATE.  Now removed this declaration.

Also removed comments which referred to dry deposition, because
this is now handled in GeosCore/mixing_mod.F90.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>