Bug fix: move QB_NUM calc in do_merra_convection to within timestep loop.
QB_NUM is the numerator of the calculation of the average mixing ratio of
the column below the cloud base. It is calculated as the sum of the
tracer concentrations (Q) weighted by grid box pressure differences
(analalous to mass).

While QB is updated within the internal timestep loop, it does not use
updated Q values since QB_NUM is calculated outside of the
internal timestep loop. Moving the calculation of QB_NUM to inside the
timestep loop causes QB to reflect updated Q values and consequently
improves mass conservation. This new method is consistent with the current
convection method in NFCLDMX.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>