Add a mass conservation test via modified ND49 diagnostic
Lizzie Lundgren has modified the diag49_mod.F so that we can print
out the total mass of a tracer (preferred: total CO2) so that we can
check how a GEOS-Chem simulation is conserving mass.

Made the following updates:

Makefile_header.mk: Added a new Makefile option MASSCONC.  If
MASSCONC=y (or MASSCONC=yes, etc.), then this will turn on the
MASSCONS C-preprocessor switch (i.e. -DMASSCONS).

GeosCore/input_mod.F: If MASSCONC=y then it will make sure that
the ND49 diagnostic is turned on, regardless of the setting in
the ND49 MENU of input.geos.  It will also reset the TRACERS array
to 1, so that we only print out masses for the first tracer (e.g.
for total CO2).  This will also make sure to turn off emissions,
regardless of the setting in the EMISSIONS MENU of input.geos.

GeosCore/diag49_mod.F: Added code to save out the total mass of
a tracer to an ASCII file after a specified number of hours.
This ASCII file will be named with the value stored in
Input_Opt%ND49_FILE.  The ND49 timeseries bpch file will be
not be created.

Right now, we only are printing out the total mass, but in the
future, we could automate this test by comparing the variance
in mass against a threshold.  Do this later...

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>