Fix typo in NXTRA definition in rrtmg_rad_transfer_mod.
Set Eastham wrote:
    (GEOS-Chem v10-01 contains) an issue in the RRTMG interface where
there is a single-character typo in SET_SPECMASK. This typo results in
dust, ozone and methane RFs being incorrectly stored - in fact it
prevents O3 and CH4 RFs from being stored at all, instead storing dust
RFs in their place (which in turn miss the last two dust bins and instead
use either sea salt or stratospheric aerosol). This is because line 1609
in rrtmg_rad_transfer_mod reads

 NXTRA=NASPECRAD-16

    When it should read

 NXTRA=NSPECRAD-16

    (note second character). This bug affects both the tropospheric and
UCX version of the code. Since NXTRA is the offset used to account for
whether or not stratospheric aerosols are inserted mid-way through the
indexing, and NASPECRAD is always 2 smaller than NSPECRAD, everything
index above 7 is set to a value 2 smaller than it should be in the mask.

    This bug occurs with or without UCX. The offset is always incorrectly
calculated, so dust, ozone and methane are never output correctly. The
easiest test is to check the ozone or methane output (they will both look
like large-diameter dust RF). Dust will look *almost* right, but will
actually include stratospheric aerosols (if UCX=yes) or sea-salt (if UCX
is not on).

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>