Fixed minor issues in the ND20 diangnostic
(1) In routine DIAG20, we modified the IF block so that we skip
computing the P(O3) and L(O3) if we are outside of the chemgrid
(i.e. where JLOP(I,J,L) == 0.

(2) Fixed a minor issue in WRITE20 -- the line of code where JFIRST
was being defined had been deleted.  Now restored.

(3) For TOMAS simulations, now avoid doing a string comparison
for the family name PSO4 on each I,J,L iteration (which is
inefficient).  Pull the test for family name into the N-loop
and store the result in a PRIVATE logical variable.

(4) Now save P(O3) as kg/m3/s and L(O3) as 1/m3/s, which are the
units that are expected by HEMCO.  This will make it easier to
convert the data files from bpch to netCDF with BPCH2COARDS.

(5) In WRITE20, now do not use parallel loops to cast from REAL*4
to REAL*8.

(6) Add a new test for in ITS_TIME_FOR_WRITE20 that will let the
rate.YYYYMMDD file be written out on the last timestep.  This lets
us use short simulations (< 1 day) for debugging.

(7) Make sure to zero all arrays after allocating.

(8) Cosmetic changes (lining up variables, etc.)

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>