Add an optional local time window to ND20 diagnostics
We now have added (commented-out) code to skip archiving the ND20
P(O3) and L(O3) diagnostics unless we are in a specified local-time
window.  Requested by Lu Shen.

Uncomment the code to activate it.  Also set the parameters
LT_START and LT_END to indicate the start and end of the local time
averaging window.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>