Add fixes for custom nested grid to tpcore_bc_mod.F
The array BC was previously used for two purposes in tpcore_bc_mod.F:
(1) for input BCs for any region and (2) for output BCs for a custom grid.
To avoid confusion, we have created an array BC_CU to be used for saving
out BCs for a custom grid. We also now use a unique LUN (IU_BC_CU) for
writing out BCs for a custom grid.

In Makefile_header.mk we now allow the user to specify NEST=cu, which will
set the -NESTED_CU CPP switch.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>