Compute ocean chemistry in all ocean boxes
Updated to allow ocean chemistry to be computed in all ocean grid boxes,
irrespective of whether there is any ice cover. If there is partial ice
cover, input of solar radiation and mercury deposition to the ocean are
reduced accordingly. This change should only affect MERRA, which contains
information on partial ice cover.