Minor fix for ISTRAT in aerosol_mod.F
The scalar ISTRAT was added to the !$OMP+PRIVATE statement for the loop
over JLOOP = 1,NTLOOP in routine RDAER (aerosol_mod.F). This caused ISTRAT
to have extremely large values when it was passed to routine GET_STRAT_OPT
and caused an out-of-bounds error in the SAD_AER array. To fix this, we
have removed ISTRAT from the !$OMP+PRIVATE statement.

We've also removed ISTRAT from another !$OMP+PRIVATE statement for the
loop over N = 1, NRHAER since it is not used in that loop.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>