Update Polar BrO chemistry
Updated with new polar BrO scheme, where BrO concentrations in the Arctic
and Antarctic are variable based on solar radiation and temperature rather
than fixed as used in previous versions of the model.
Parameterization makes [BrO] a simple linear function of T. An upper threshold on
how much ice is present was also added. The imposed diurnal cycle for
Br/BrO was moved so that it applies only to the daily mean concentrations
from p-TOMCAT, not to the instantaneous values calculated every hour for
the polar BrO scheme. Relevant diagnostics also added.