Add new routine GIGC_ERROR to error_mod.F
Lizzie has added a new subroutine GIGC_ERROR to GeosUtil/error_mod.F. This
routine sets RC to GIGC_FAILURE and prints a message and location to the
log file. An example of how to use this routine is provided below:

   ! Verify correct initial units
   IF ( TRIM( State_Chm%Trac_Units ) /= 'kg/kg dry' ) THEN
      MSG = 'Incorrect initial units:' // TRIM( State_Chm%Trac_Units )
      LOC = 'UNITCONV_MOD: Convert_DryKgKg_to_MoistKgKg'
      CALL GIGC_Error( MSG, RC, LOC)
   ENDIF

In main.F, you can then add a check for RC and stop GEOS-Chem if there's an
error. For example:

   IF ( RC /= GIGC_SUCCESS ) THEN
      CALL ERROR_STOP('Unit conv error before AIRQNT', 'main.F')
   ENDIF

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>