Add Arctic Rivers
Added input of Hg to Arctic Ocean from circumpolar rivers with
concentrations matching Fisher et al. (2012) for a total flux
of 95 Mg/a (average). Also added interannual variability of the
river flow inputs (not concentration) based on Fisher et al. (2013).
This uses a precomputed monthly scaling factor for each month in
May-October and each year in 1979-2009, calculated relative to the
mean over this period. Years after 2009 use the mean value.