Tie photo-reducible HgII(aq) to UV-B radiation
Tied the pool of HgII that is reducible by photo-reduction to available
UV radiation (scaled by the ratio of a given year's UV to 2006-2010 UV
- both based on TOMS O3 columns for a given gridbox and month). Missing
O3 data are replaced with climatological mean for the same box/month.
The default value for the photo-reducible fraction is 0.4 for consistency
with standard simulations. For more info, see code or Fisher et al. (2013).