Add comments stating we convert MERRA2 pressure quantities to hPa
In prior met fields, PS, TROPPT, SLP were in hPa, but now in MERRA2
we save those to disk as Pa.  Therefore, we convert to hPa when
reading these from disk.  Add comments stating this to a couple of
routines.

NOTE: We are doing this for backwards compatibility, but eventually
we will recode GEOS-Chem to use pressure quantities in Pa.  This will
make it easier to interface with the GEOS-5 GCM.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>