Bug fix: set XMW = 136e-3 for SOA species MTPO in drydep_mod.F
The previous code had XMW = 136e-3 kg/mol for MTPO.  The new species
database sets XMW to 136.23e-3 kg/mol.  Therefore, there is a minor
discrepancy that will cause the code to fail difference tests.

For now, add a kluge to set XMW = 136e-3 for MTPO, so that we can
replicate the prior behavior.  Also added a note that we should update
this to 136.2e-3 kg/mol soon.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>