Additional fixes to resolve differences in Hg simulation when using mp vs sp
Now pass a scalar DEP_HG to subroutine ADD_Hg2_SNOWPACK in wetscav_mod.F.
We previously passed DSTT(NN,1,I,J), but this was causing differences in
the geos5_4x5_Hg unit test simulation in the HG-SRCE diagnostic for tracer
21 (Hg2/HgP deposition to snow/ice).

NOTE: We are still getting differences in the geos5_4x5_Unit Teste. The
causes of these differences have been identified, but a solution has still
not been found. The remaining differences are:
  (1) HG-SRCE for tracer 21 - Tracked to the call to ADD_Hg2_SNOWPACK
      for WET_Hg2 in convection_mod.F (currently lines 1840-1841).
  (2) PL-HG2-$ for tracer 4 - Tracked to the parallel DO loop in
      CHEMMERCURY where we modify DEP_HG2G_SALT and eventually add it to
      AD03_Hg2_SS. When we comment out the !$OMP directives for this loop,
      the differences in PL-HG2-$ go away.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>