Restore DO loop in mercury_mod.F to the original order
We had changed a DO loop in CHEMMERCURY from L, J, I to the proper order
I, J, L  to be more efficient. However, this change caused differences in
the PL-HG2-$ diagnostic for tracer 4 (stored in array AD03_Hg2_SS) when
using mp vs sp. The reasons for these differences are not known, so we
have restored the loop to its original order for now.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>