Optimization bug fix in Hg code in convection_mod routine NFCLDMX.
Consolidated the calculation of WET_Hg2 in NFCLDMX from two to one lines.
Doing this removed an instance where WET_Hg2 appears on both the lhs and
rhs of an assignment and solved an optimization bug in the mercury
simulation. Prior to this fix, geos5_4x5_Hg runs did not pass the unit
test due to differences in HG-SRCE for tracer 21 that were tracked
to the call to ADD_Hg2_SNOWPACK occurring directly after the lines
altered for this bug fix. It is unclear why the previous code caused
this optimization problem.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>