Add WD_LiqAndGas and WD_ConvFactor fields to species database
The fields WD_LiqAndGas and WD_ConvFactor allow us to generalize many
of the specific IF blocks in COMPUTE_F and RAINOUT.  This will let us
collapsed down much of these routines.

Also note: there is some confusion about how the suppression of
scavenging is applied in RAINOUT.  The RETURN statements may actually
be preventing the SO2 block from occurring.  Need to investigate
this further.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>