Update nested grid region definitions for MERRA2
We define the following regions for the MERRA2 nested grids:

   CH : [ -11,   60, 55, 150 ]
   EU : [  30,  -30, 70,  50 ]
   NA : [  10, -140, 70, -40 ]

The regions are the same as the GEOS-5 regions since the horizontal
resolutions for GEOS-5 and MERRA2 are similar (0.5x0.666 vs 0.5x0.625).
The only difference is that the CH grid has been extended eastward 10
degrees to include all of India.

The I1_BC_*,J1_BC_*,I2_BC*,J2_BC* indicies for MERRA2 in tpcore_bc_mod.F
have been updated accordingly. The IIPAR and JJPAR values for the MERRA2
nested grids in CMN_SIZE_mod.F have also been updated.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>