Remove MAXDEP from drydep_mod.F; use NUMDEP instead
Module drydep_mod.F used to use MAXDEP to allocate various arrays.
This was a holdover from when the dry deposition code used old
F77-style common blocks.

With the species database, we now know the number of dry-depositing
species before we even call INIT_DRYDEP. Therefore, we can use
NUMDEP (which is the same as State_Chm%nDryDep) to allocate arrays.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>