Now keep track of actually wet-deposited species
The State_Chm%nWetDep field represents the number of species that
can undergo wet deposition.  But if wet deposition is turned off,
then State_Chm%nWetDep is unchanged.

Therefore, we use the variable NSOL in wetscav_mod.F to denote the
actual number of wet-deposited species.  Restore routine
GET_WETDEP_NSOL to wetscav_mod.F to pass this value to various
diagnostic routines.

Also, in INIT_WETSCAV, we only update the IDWETD array if wet deposition
or convection is turned on. This should help to avoid out-of-bounds
errors.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>