Use State_Met variable SWGDN for incident radiation to distinguish from net
State_Met%RADSWG is net radiation at ground for all MET fields except
GEOS-FP and MERRA2. Previously, incident radiation in GEOS-FP and MERRA2
was assigned to RADSWG despite the definition difference. Now there is a
new State_Met variable, SWGDN, that is set for the incident radiation.
This includes GEOS-FP, MERRA2, and also MERRA.

Where RADSWG is used, there are now pre-processor switches that use
either RADSWG or SWGDN. SWGDN is always used if GEOS-FP or MERRA2.
If using MERRA, RADSWG is still used to preserve legacy methods except
in parts of the mercury code where incident radiation is preferred over
net radiation and therefore SWGDN is used.

In HEMCO, the variable ExtState%RADSWG is set equal to State_Met%SWGDN if
using GEOS-FP or MERRA2, and set equal to State_Met%RADSWG otherwise.
Using a single ExtState variable is the simplest approach and can be
changed in the future if developers wish to have both net and incident
radiation available.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>