Now use the species database to initialize quantities in dust_mod.F
We now initialize the DRY* flags, as well as the values of
Input_Opt%DUSTDEN and Input_Opt%DUSTREFF (these are used in both
sulfate_mod.F and aerosol_mod.F) with the values from the species
database.

NOTE: We did not touch the TOMAS code in dust_mod.F.  Also, we will
need to modify the species database to zero out drydep and wetdep
ID's when drydep/wetdep are shut off.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>