Now define temperature-dependent rainout efficiencies
The species database field WD_RainoutEff is now a 3-element vector:

  Element 1: Rainout efficiency when T < 237 K.
  Element 2: Rainout efficiency when 237 K <= T < 258 K.
  Element 3: Rainout efficiency when T > 258 I.

This allows us to apply Qiaoqiao Wang's impaction scavenging algorithm:
Namely, BC, dust, and HNO3 tracers as IN, and allows them to be
scavenged below 258 K... and then to suppress scavenging for other
aerosol tracers (and SO2) when 237 K <= T < 258 K.

The prior code used a confusing IF ( N = IDTxxx ) THEN ... block
to do this.  We have replicated this functionality without having
to resort to this type of struture.

Bob Yantosca validated these updates with an aerosol-only difference
test w/r/t GEOS-Chem v11-01c.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>