Define temperature-dependent scale factors for KC in the species database
The impaction scavenging modification to the wet deposition involves
multiplying the KC (cloud condensation -> precipitation) rate by
temperature-dependent scale factors.  These were implemented in an
ugly IF block in the routine F_AEROSOL.  The logic of the IF block
was hard to penetrate.

Therefore, we have added a 3-element vector field WD_KcScaleFac to
the species database.  (This is conceptually similar to WD_RainoutEff.)
WD_KcScaleFac allows you to specify a scale factor to modify the KC
rate for 3 temperature regimes (1) T < 237 K; (2) 237 K <= T < 258 K,
and (3) T >= 258 K.

NOTE: Commit these changes now but need to debug the code further
to avoid a floating overflow.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>