For now, leave rainout turned on for HgP when 237 <= T < 258K
In order to replicate the prior results, we have to NOT shut off
rainout for HgP aerosol when 237 <= T < 258 K.  But we also have to
leave the trap that multiplies Kc by 0.5.

Consult with Team Hg before modifying the code further.  For now,
let's be content that we replicated the prior results.

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>