Now specify input and output ocean Hg restart file names in input.geos
We previously specified one filename for the ocean restart file and
assumed that file was located in the run directory. However, we now give
users the option to get the initial ocean restart from the
{DATA_ROOT}/GC_RESTARTS/ path. To allow for users to save out ocean
restart files in their run directory, we added options for input and
output ocean Hg restart files in the input.geos files.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>