Now use WD_Is_SO2 and WD_Is_HNO3 to flag SO2 and HNO3 wet deposition
In routines COMPUTE_F, RAINOUT, and WASHOUT, we now use the WD_Is_HNO3
and WD_Is_SO2 fields of the species database object to determine if
we have to do special handling for HNO3 and SO2 wet scavenging
computations.  This replaces the old "IF ( N == IDTHNO3 ) THEN" and
"IF ( N == IDTSO2 ) THEN" tests.

Also added some comments to the !REMARKS section of routine Spc_Create
(in Headers/species_mod.F90).

Signed-off-by: Bob Yantosca <yantosca@seas.harvard.edu>