Avoid differences caused by initialization error in KPP
We encountered differences between single processor and multi-processor
runs when KPP was turned on. A discussion post on the GEOS-Chem wiki
attributes the differences to an initialization issue in KPP and recommends
setting unassigned V() elements to 1d-99.  See:

  http://wiki.seas.harvard.edu/geos-chem/index.php/Talk:Interfacing_GEOS-Chem_with_KPP

In this case, the species CO2 and DRYDEP (index numbers vary) were not
initialized in each of the five chemistry mechanisms. These species are
both set to dead in globchem.dat, so setting them to 1d-99 is fine.

Signed-off-by: Melissa Sulprizio <mpayer@seas.harvard.edu>