Fix indexes in offline dust calculation when UCX is on.
Seb Eastham writes:

"On or around line 1450 of dust_mod.F [in v10-01], the dry radius and
extinction of dust are retrieved from QQAA and RDAA. If RRTMG is active,
the single scattering albedo and asymmetry factors are also retrieved,
from SSAA and ASYMAA. However, they are always read from entry [X,Y,Z,6]
of the corresponding array.

Unfortunately, this index is only correct if the UCX is off; if the UCX
is on, the entry should be 8, not 6 (based on fast_jx_mod, where these
arrays are populated). It looks like an index variable was prepared for
this purpose but it's not used. As it is, when the UCX is on, dust radius,
surface area and scattering will actually be calculated based on
stratospheric sulfate aerosol properties."

The updated implementation is based on the implementation in rdust_online.

Signed-off-by: Lizzie Lundgren <elundgren@seas.harvard.edu>